requirements.txt
mystmd
jupyterlab-myst
pandas
pyyaml
rich
ipython